requirements.txt
discord.py
PyNaCl==1.3.0
async-timeout==3.0.1
dnspython==1.16.0